requirements.txt
flask
gunicorn
segno
Pillow